added methods for category lookups

+++ Gettext/GettextLib/GettextPseudo.cs
@@ -7,7 +7,7 @@ namespace GettextLib
             return Utils.PseudoTranslate(base._(msgid));
         }
 
-        public override string NGettext(string msgid, string msgidPlural, int n)
+        public override string NGettext(string msgid, string msgidPlural, long n)
         {
             return Utils.PseudoTranslate(base.NGettext(msgid, msgidPlural, n));
         }
